refactor(패키지 추가) : MVC 패턴 기반 패키지 추가
 - domain 패키지 추가
 - view 패키지 추가
