+++ src/main/java/carracing/MainApplication.java
@@ -1,5 +1,7 @@
 package carracing;
 
+import carracing.domain.CarRacingProgram;
+
 public class MainApplication {
 
     public static void main(String[] args) {
