+++ src/main/java/carracing/ui/CarRacingProgramPhrase.java
@@ -1,4 +1,4 @@
-package carracing;
+package carracing.ui;
 
 public class CarRacingProgramPhrase {
