+++ src/test/java/carracing/CarTest.java
@@ -1,5 +1,6 @@
 package carracing;
 
+import carracing.domain.Car;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
